local/requirements.txt
# PDF Translator Local v2
# Tesseract OCR + OpenAI + PyQt5

# GUI
PyQt5>=5.15.0

# PDF Processing
PyMuPDF>=1.23.0

# OCR
pytesseract>=0.3.10

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Translation
openai>=1.0.0
